Split IntUiBridge into separate files according to components (#670)
* extract IntUiBridgeCheckbox


* extract IntUiBridgeRadioButton


* extract IntUiBridgeCircularProgress


* extract IntUiBridgeButton


* extract IntUiBridgeChip


* extract IntUiBridgeComboBox


* extract IntUiBridgeDropDown


* extract IntUiBridgeSimpleListItem


* iterate on IntUiBridgeProgress


* iterate on IntUiBridgeLink


* iterate on IntUiBridgeMenu


* iterate on IntUiBridgePopupContainer


* iterate on IntUiBridgeSegmentedControl


* iterate on IntUiBridgeSlider


* iterate on IntUiBridgeTextArea


* iterate on IntUiBridgeTextField


* iterate on IntUiBridgeLazyTree


* iterate on IntUiBridgeTab


* iterate on IntUiBridgeButton


* extract on IntUiBridgeTooltip


* extract on IntUiBridgeGroupHeader


* extract on IntUiBridgeSelectableLazyColumn


* extract on IntUiBridgeDivider


* extract on IntUiBridgeText


* reformat all the IntUiBridge* files


---------
