+++ ide-laf-bridge/api/ide-laf-bridge.api
@@ -150,7 +150,7 @@ public final class org/jetbrains/jewel/bridge/theme/BridgeThemeColorPaletteKt {
 	public static final fun readFromLaF (Lorg/jetbrains/jewel/foundation/theme/ThemeColorPalette$Companion;)Lorg/jetbrains/jewel/foundation/theme/ThemeColorPalette;
 }
 
-public final class org/jetbrains/jewel/bridge/theme/IntUiBridgeKt {
+public final class org/jetbrains/jewel/bridge/theme/IntUiBridgeTextKt {
 	public static final fun retrieveConsoleTextStyle ()Landroidx/compose/ui/text/TextStyle;
 	public static final fun retrieveDefaultTextStyle ()Landroidx/compose/ui/text/TextStyle;
 	public static final fun retrieveEditorTextStyle ()Landroidx/compose/ui/text/TextStyle;
